:date: Introduce a basic 'add to calendar' functionality
WHenever the user sets a date in the future, she or he will be given the opportunity to create a calendar event based on the chosen date.

This is going to be expended upon but this first iteration introduces this possibility to the 'discover moment' dialog. By setting a value to any of the inpur fields, a 'Add to calendar' link will appear.

The link and the options that show after clicking on it aren't styled at this point, but will be.