:zap: Introduce 'AddToCalendar' as a component
- Update 'DateCard' so that when the date is in the future, the user is able to add it as an event to a calendar
- All modes support this change

Because of the nature of dates, we have to use 'nowDate' to establish if the date passed to 'DateCard' is in the future or not. This differs from the implementation on 'DiscoverMoment' component, that uses 'nowDailyDate' instead, for performance reasons.

Also update the testing setup so we mock 'navigator' correctly while in the test environment.